Finished obstacle collison + other
Finished obstacle collision, changed Bounds3D to accurately represent the top of a 3D object. Did some work on buildings.

Would probably still be a good idea to overhaul how buildings work to clean up the sudden spawn and the weird block of adding to the array list in the Environment class